+++ HISTORY.rst
@@ -8,6 +8,7 @@
 - Update the Debian Docker image from version 11.5-slim to 11.6-slim.
 - Ensure the workflow state action column is not shown for the workflow
   state runtime proxies where is does not make sense to show.
+- Add escalation list column to workflow states list view.
 
 4.4.4 (2023-02-14)
 ==================